Homework #3, Исправил копипсту другим способом
